+++ lib/src/credentials.dart
@@ -88,7 +88,12 @@ class EmailPasswordAuthProvider {
 
   /// Completes the reset password procedure by providing the desired new [password] using the 
   /// password reset [token] and [tokenId] that were emailed to a user.
-  Future<void> resetPassword(String password, String token, String tokenId) {
-    return realmCore.emailPasswordResetPassword(application, password, token, tokenId);
+  Future<void> completeResetPassword(String password, String token, String tokenId) {
+    return realmCore.emailPasswordCompleteResetPassword(application, password, token, tokenId);
+  }
+
+  /// Sends a password reset [email] to initiate the password reset procedure.
+  Future<void> resetPassword(String email) {
+    return realmCore.emailPasswordResetPassword(application, email);
   }
 }