+++ lib/src/native/realm_core.dart
@@ -951,7 +951,7 @@ class _RealmCore {
     return completer.future;
   }
 
-  Future<void> emailPasswordResetPassword(Application application, String password, String token, String tokenId) {
+  Future<void> emailPasswordCompleteResetPassword(Application application, String password, String token, String tokenId) {
     final completer = Completer<void>();
     using((arena) {
       _realmLib.invokeGetBool(() => _realmLib.realm_app_email_password_provider_client_reset_password(
@@ -959,6 +959,20 @@ class _RealmCore {
             password.toRealmString(arena).ref,
             token.toUtf8Ptr(arena),
             tokenId.toUtf8Ptr(arena),
+            Pointer.fromFunction(void_completion_callback),
+            completer.toPersistentHandle(),
+            _deletePersistentHandleFuncPtr,
+          ));
+    });
+    return completer.future;
+  }
+
+  Future<void> emailPasswordResetPassword(Application application, String email) {
+    final completer = Completer<void>();
+    using((arena) {
+      _realmLib.invokeGetBool(() => _realmLib.realm_app_email_password_provider_client_send_reset_password_email(
+            application.handle._pointer,
+            email.toUtf8Ptr(arena),
             Pointer.fromFunction(void_completion_callback),
             completer.toPersistentHandle(),
             _deletePersistentHandleFuncPtr,